requirements.txt
openai
python-dotenv
PyGithub
requests
